Move python files to src directory and add pyproject.toml
