+++ .gitignore
@@ -2,3 +2,9 @@ Future Ideas and Issues to Address.md
 README.md
 help_future.md
 help.md
+*.log
+__pycache__/
+*.py[cod]
+venv/
+lib.lib
+*.yml
