--- beta/yt_urls.yml
@@ -1,49 +1,0 @@
-- https://www.youtube.com/watch?v=M5yblXlUk9M
-- https://www.youtube.com/watch?v=a06Dz3iDz9o
-- https://www.youtube.com/watch?v=Aj1ZO28Fa9I
-- https://www.youtube.com/watch?v=a-cC5KUhPbw
-- https://www.youtube.com/watch?v=yO5GaW3F30o
-- https://www.youtube.com/watch?v=Xsj_fqOTs5U
-- https://www.youtube.com/watch?v=oh6_6-X-cGk
-- https://www.youtube.com/watch?v=vF3KH99IdHs
-- https://www.youtube.com/watch?v=iQQCpDLQOq4
-- https://www.youtube.com/watch?v=Cc826uB1m1Y
-- https://www.youtube.com/watch?v=x7mJWBumits
-- https://www.youtube.com/watch?v=urf81Olp7c8
-- https://www.youtube.com/watch?v=OiUNEMtZS6g
-- https://www.youtube.com/watch?v=Gqw8D0nLbqo
-- https://www.youtube.com/watch?v=nSkjyT3PDSU
-- https://www.youtube.com/watch?v=jbaMPy02SUA
-- https://www.youtube.com/watch?v=BQ35OEQhjfc
-- https://www.youtube.com/watch?v=YN3hzchcw5k
-- https://www.youtube.com/watch?v=SOMZHqPM7eY
-- https://www.youtube.com/watch?v=TSLN4x8ow4I
-- https://www.youtube.com/watch?v=m3TVPah4aVY
-- https://www.youtube.com/watch?v=LzPdy5lC_Hk
-- https://www.youtube.com/watch?v=hL9eq-TVwyk
-- https://www.youtube.com/watch?v=wMzzInTxiOo
-- https://www.youtube.com/watch?v=2uuWt40-gQ4
-- https://www.youtube.com/watch?v=pjGMX6Ivcks
-- https://www.youtube.com/watch?v=mMMPxPJVeXY
-- https://www.youtube.com/watch?v=GyEDjmfN6Ys
-- https://www.youtube.com/watch?v=pSwdXprCoB4
-- https://www.youtube.com/watch?v=0JKG5VIiESY
-- https://www.youtube.com/watch?v=H7UeTkowQZE
-- https://www.youtube.com/watch?v=f_ozolV7V1M
-- https://www.youtube.com/watch?v=uCs-9pf4mTg
-- https://www.youtube.com/watch?v=G9sENKMK8Tk
-- https://www.youtube.com/watch?v=n_Snq10mgPk
-- https://www.youtube.com/watch?v=_CAS6e_BRvo
-- https://www.youtube.com/watch?v=YaqaNj90u3k
-- https://www.youtube.com/watch?v=DujEvJIW_9I
-- https://www.youtube.com/watch?v=0qWbdsU1kFo
-- https://www.youtube.com/watch?v=z_ipM46uW_Y
-- https://www.youtube.com/watch?v=M5zIMekrOPU
-- https://www.youtube.com/watch?v=_T5hu6sPB8I
-- https://www.youtube.com/watch?v=lI9MV5rPUhg
-- https://www.youtube.com/watch?v=kj3V-_yeya4
-- https://www.youtube.com/watch?v=YN-O_A6rr3I
-- https://www.youtube.com/watch?v=SaFKOuR0VAo
-- https://www.youtube.com/watch?v=3uUkvMCYlYs&list=RDGMEMHDXYb1_DDSgDsobPsOFxpA&index=2
-- https://www.youtube.com/watch?v=WavXAhu4HWc&list=OLAK5uy_nv6WNWUTzobgNKAcHj6PwgiJZKdgKzab8
-- https://www.youtube.com/watch?v=2qYLxw0Knu4&list=RDGMEMHDXYb1_DDSgDsobPsOFxpA&start_radio=1&rv=amAIoxY56-0
