+++ pyproject.toml
@@ -1,0 +1,34 @@
+[tool.poetry]
+name = "mariana-music-player"
+version = "0.4.2"
+description = "Best Music Player Ever - CLI"
+authors = ["Your Name <you@example.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+colored = "^1.4.3"
+comtypes = "^1.1.11"
+pyPodcastParser = "^2.0.0"
+numpy = "^1.22.4"
+Pillow = "^9.1.1"
+praw = "^7.6.0"
+PyDrive = "^1.3.1"
+pygame = "^2.1.2"
+PyInquirer = "^1.0.3"
+pynput = "^1.7.6"
+python-vlc = "^3.0.16120"
+PyYAML = "^6.0"
+regex = "^2022.6.2"
+"ruamel.yaml" = "^0.17.21"
+shazamio = "^0.1.1"
+sounddevice = "^0.4.4"
+tabulate = "^0.8.9"
+tqdm = "^4.64.0"
+toml = "^0.10.2"
+youtube_dl = "^2021.12.17"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
