--- requirements.txt
@@ -1,17 +1,0 @@
-pygame
-youtube_dl
-tabulate
-numpy
-ruamel.yaml
-pydrive
-python-vlc
-regex
-PyInquirer
-pynput
-comtypes
-praw
-shazamio
-colored
-pycaw
-# librosa (optional)
-# Doesn't work, so commented: git+https://github.com/Vivojay/pafy@develop
