+++ src/lyrics_provider/lyrics_view_window.py
